+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Test and beta releases will have experimental functions enabled by default.
 
 * Rewritten URI restrictions for better compatibility
 * Simplified restricting / allowing contacts
+* Added restriction for browser downloads
 * Updated traditional Chinese translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)